--- src/components/Fieldset/index.js
@@ -1,5 +1,0 @@
-import CategoryButton from './CategoryButton/CategoryButton.jsx';
-import CategoryFieldset from './Fieldset/Fieldset.jsx';
-import TextField from './TextField/TextField';
-
-export { CategoryButton, CategoryFieldset, TextField };
