--- src/components/Filter/index.js
@@ -1,2 +1,0 @@
-export { default as Filter } from './Filter';
-export { default as Filters } from './Filters';
