--- src/components/Input/index.js
@@ -1,2 +1,0 @@
-export { default as InputItem } from './InputItem/InputItem';
-export { default as InputList } from './InputList/InputList';
