--- src/components/ReviewContentItem/index.js
@@ -1,1 +1,0 @@
-export { default as ReviewContentItem } from './ReviewContentItem';
