--- src/components/ReviewDownload/index.js
@@ -1,1 +1,0 @@
-export { default as ReviewDownload } from './ReviewDownload';
