+++ src/components/index.js
@@ -1,13 +1,7 @@
 export * from './BoardCard';
 export * from './Card';
 export * from './Carousel';
-export * from './Filter';
-export * from './Fieldset';
-export * from './Filter';
-export * from './Input';
 export * from './ListHeader';
 export * from './MainPageListItem';
 export * from './PopUp';
-export * from './ReviewContentItem';
-export * from './ReviewDownload';
 export * from './Search';
