--- src/constants/examReview.js
@@ -1,1 +1,0 @@
-export const FILE_MAX_SIZE = 1024 * 1024 * 10;
