+++ src/constants/index.js
@@ -1,7 +1,4 @@
 export * from './about.js';
 export * from './admin.js';
-export * from './align.js';
-export * from './dropDown.js';
-export * from './examReview.js';
 export * from './search.js';
 export * from './seconds.js';
