+++ src/feature/exam/component/index.js
@@ -1,0 +1,9 @@
+export { default as CategoryButton } from './CategoryButton/CategoryButton';
+export { default as CategoryFieldset } from './Fieldset/Fieldset';
+export { default as Filter } from './Filter/Filter';
+export { default as Filters } from './Filters/Filters';
+export { default as InputItem } from './InputItem/InputItem';
+export { default as InputList } from './InputList/InputList';
+export { default as ReviewContentItem } from './ReviewContentItem/ReviewContentItem';
+export { default as ReviewDownload } from './ReviewDownload/ReviewDownload';
+export { default as TextField } from './TextField/TextField';
