+++ src/feature/exam/constant/index.js
@@ -1,0 +1,2 @@
+export * from './align.js';
+export * from './exam.js';
