+++ src/feature/exam/lib/index.js
@@ -1,0 +1,2 @@
+export * from './convertDS.js';
+export { default as CheckExamPeriodRoute } from './CheckExamPeriodRoute';
