+++ src/pages/ExamReviewsPage/ExamReviewsPage.jsx
@@ -1,14 +1,17 @@
 import { Link } from 'react-router-dom';
 import { useQuery } from '@tanstack/react-query';
 
+import { getNoticeLine } from '@/apis';
+
 import { useScrollRestoration } from '@/shared/hook';
 import { AppBar, Icon, WriteButton } from '@/shared/component';
 import { QUERY_KEY, STALE_TIME } from '@/shared/constant';
 
-import { getNoticeLine } from '@/apis';
+import { Filter, Filters } from '@/feature/exam/component';
+import { YEARS, SEMESTERS, EXAM_TYPES } from '@/feature/exam/constant';
+
 import { SearchExamReviewsSuspense } from '@/pages/ExamReviewsPage';
-import { Search, Filter, Filters } from '@/components';
-import { YEARS, SEMESTERS, EXAM_TYPES } from '@/constants';
+import { Search } from '@/components';
 
 import styles from './ExamReviewsPage.module.css';
